Create QUINE
"QUINE" is a SWIPL expert system using automated, cyclic port forensics along with expert system components of artificial intelligence.  This script was developed for partial completion of an MSc. in Cyber Security from Utica College, and uses a "Best-First Search" algorithm for finding vulnerabilities associated with port and protocols.

From the SWI-Prolog Manual:

2.10.2.1 Using PrologScript

A Prolog source file can be used directly as a Unix program using the Unix #! magic start. The same mechanism is useful for specifying additional parameters for running a Prolog file on Windows. The Unix #! magic is allowed because if the first letter of a Prolog file is #, the first line is treated as a comment.11 To create a Prolog script, use one of the two alternatives below as first line. The first can be used to bind a script to a specific Prolog installation, while the latter uses the default prolog installed in $PATH.

\verb$#!/path/to/swipl$
\verb$#!/usr/bin/env swipl$

2.10.2.2 Creating a shell script

With the introduction of PrologScript (see section 2.10.2.1), using shell scripts as explained in this section has become redundant for most applications.

Especially on Unix systems and not-too-large applications, writing a shell script that simply loads your application and calls the entry point is often a good choice. A skeleton for the script is given below, followed by the Prolog code to obtain the program arguments.

#!/bin/sh

base=<absolute-path-to-source>
PL=swipl

exec $PL -q -f "$base/load" --
:- initialization go.

go :-
        current_prolog_flag(argv, Arguments),
        go(Arguments).

go(Args) :-
        ...