Add files via upload
Please check out the new guides for the decoder-shell (Dshell) framework: the Dshell User Guide and the Dshell Developer Guide.
The Dshell User Guide is a general user guide which details installation and both basic and advanced analysis usage with examples.
The Dshell Developer Guide is a guide to plugin development which provides basic examples, core function and class definitions, and an overview of data flow. This guide will help end users develop new, custom plugins as well as modify existing plugins.